# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Simple CRUD",
    "description": "CRUD per la gestione degli utenti.",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {

    "/users/list/{page}": {
      "get": {
        "operationId": "usersPaginate",
        "tags": ["/users"],
        "description": "Ritorna una pagina di 5 di utenti ",
        "parameters": [{
            "name": "page",
            "in": "path",
            "description": "numero di pagina",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },

    "/users/add": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Aggiunge un utente",
        "description": "",
        "operationId": "addUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "utente da inserire",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid Order"
          }
        }
      }
    }



  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "name": {
          "type": "object",
          "required": [
            "first",
            "last"
          ],
          "properties": {
            "first": {
              "type": "string",
              "description": "nome dell'utente"
            },
            "last": {
              "type": "string",
              "description": "cognome dell'utente"
            }
          }
        },
        "email": {
          "type": "string",
          "format": "email",
          "pattern": "/^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$/",
          "description": "indirizzo e-mail dell'utente"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "data di creazione dell'utente"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "data di ultima modifica dell'utente"
        }
      },
      "xml": {
        "name": "User"
      }
    }
  }
}